rename StratumNode to CategoryNode, part 2
