+++ client/lib/CategoryNode/getOrigin.js
