+++ client/lib/CategoryNode/getRadius.js
