+++ client/lib/CategoryNode/getScale.js
