+++ client/lib/CategoryNode/index.js
