+++ client/lib/CategoryNode/makeClosed.js
