+++ client/lib/CategoryNode/makeOpened.js
