+++ client/lib/CategoryNode/remove.js
