+++ client/lib/CategoryNode/render.js
