+++ client/lib/CategoryNode/setLoadingAnimation.js
