+++ client/lib/CategoryNode/stratumNode.css
