+++ client/lib/CategoryNode/template.js
